.r Extract Path type to its own file (go test green; automated symbol move)
internal/injection/render_injection_bundle.go used to declare the
Path string-wrapper type and its methods at the BOTTOM of a 1937-line
monolith, after every function that consumed it.  A reader skimming
the package met the abstraction last instead of first.

Move the type + its five methods (String, Parent, Join, Base, IsDir)
into a new internal/injection/path.go.  Zero behavior change.  Also
drop the empty init() placeholder that lived next to the old
declaration; it had no body.
