+++ internal/injection/render_injection_bundle.go
@@ -1934,29 +1934,4 @@ func containsPath(stack []Path, candidate Path) bool {
 	return false
 }
 
-type Path string
-
-func (p Path) String() string {
-	return string(p)
-}
-
-func (p Path) Parent() Path {
-	return Path(filepath.Dir(string(p)))
-}
-
-func (p Path) Join(rel string) Path {
-	return Path(filepath.Join(string(p), rel))
-}
-
-func (p Path) Base() string {
-	return filepath.Base(string(p))
-}
-
-func (p Path) IsDir() bool {
-	info, err := os.Stat(string(p))
-	return err == nil && info.IsDir()
-}
-
-func init() {
-	// Keep the regex-free helpers isolated; no init-time side effects.
-}
+// Path lives in path.go.
